Rename Buildvana.Tool.Commands to Buildvana.Tool.Subcommands
The Commands namespace shadowed LibGit2Sharp.Commands in GitService (worked
around with a GitCommands alias) and read awkwardly beside the rest of the
command infrastructure. "Subcommands" is unambiguous and more accurate: clean,
build, restore, etc. are subcommands of bv. Rename the namespace and drop the
now-unnecessary GitCommands alias, restoring the plain Commands.Stage call.
