+++ src/Buildvana.Tool/Program.cs
@@ -10,7 +10,6 @@ using Buildvana.Core.HomeDirectory;
 using Buildvana.Core.Json;
 using Buildvana.Core.Process;
 using Buildvana.Tool.CommandLine;
-using Buildvana.Tool.Commands;
 using Buildvana.Tool.Configuration;
 using Buildvana.Tool.Infrastructure.Execution;
 using Buildvana.Tool.Infrastructure.Logging;
@@ -20,6 +19,7 @@ using Buildvana.Tool.Services.PublicApiFiles;
 using Buildvana.Tool.Services.ServerAdapters;
 using Buildvana.Tool.Services.Solution;
 using Buildvana.Tool.Services.Versioning;
+using Buildvana.Tool.Subcommands;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Spectre.Console;
