+++ src/Buildvana.Tool/Subcommands/TestCommand.cs
@@ -7,7 +7,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Buildvana.Tool.Infrastructure.Execution;
 
-namespace Buildvana.Tool.Commands;
+namespace Buildvana.Tool.Subcommands;
 
 [ImplementsCommand("test", consumesAllArguments: true)]
 [Description("Build all projects and run tests.")]
